refactoring: Title and paginagion blocks from People and Colors pages were moved to own layer
