+++ src/layouts/TitleWithPagination/TitleWithPagination.module.scss
@@ -1,0 +1,26 @@
+.page {
+  font-family: Fira Sans, sans-serif;
+  color: var(--primary-text-color);
+  display: flex;
+  flex-direction: column;
+}
+
+.title {
+  font-weight: 600;
+  font-size: 2.5rem;
+  margin-bottom: 0.6rem;
+}
+
+.description {
+  font-size: 0.9rem;
+}
+
+.children {
+  display: flex;
+  flex-wrap: wrap;
+  margin-top: 1.4rem;
+}
+
+.pagination {
+  align-self: flex-end;
+}
