--- src/screens/Colors/Colors.module.scss
@@ -1,22 +1,0 @@
-.colors-page {
-  font-family: Fira Sans, sans-serif;
-  color: var(--primary-text-color);
-  display: flex;
-  flex-direction: column;
-
-  &__title {
-    font-weight: 600;
-    font-size: 2.5rem;
-  }
-
-  &__colors {
-    display: flex;
-    flex-direction: row;
-    flex-wrap: wrap;
-    margin-top: 1.4rem;
-  }
-
-  &__pagination {
-    align-self: end;
-  }
-}
