--- src/screens/People/People.module.scss
@@ -1,31 +1,0 @@
-.users-page {
-  font-family: Fira Sans, sans-serif;
-  color: var(--primary-text-color);
-  display: flex;
-  flex-direction: column;
-
-  &__title {
-    font-weight: 600;
-    font-size: 2.5rem;
-    margin-bottom: 0.6rem;
-  }
-
-  &__description {
-    font-size: 0.9rem;
-  }
-
-  &__users {
-    display: flex;
-    flex-wrap: wrap;
-    margin-top: 1.4rem;
-  }
-
-  &__user {
-    margin-right: 2rem;
-    margin-bottom: 2rem;
-  }
-
-  &__pagination-wrapper {
-    align-self: flex-end;
-  }
-}
